+++ docs/changelog.rst
@@ -18,7 +18,8 @@ Chula v0.13.0 (dev)
 *Still under development*
 
 - Add support for using the Eventlet wsgi provider.
-- Add max-request and access-log support to :command:`chula-run`.
+- Add max-request, keep-alive, preload, worker-provider, and
+  access-log support to :command:`chula-run`.
 - Switch to httplib2 when running bat tests.
 - Fix various http GET/POST bugs.